requirements.txt
python-mpv
iterfzf
